Rework archive keyring generation and verification
We generate the archive keyring from all the individual certificates,
to make extending them easier.

We switch from signing an md5sums file to simply signing the individual
certificate files, so that this permits renames without invalidating
the signatures, and where the signature will be based on a safer digest
function anyway.

To perform the verification of the individual certificates, we use the
Stateless OpenPGP (SOP) CLI which is a draft specification proposed for
adoption by the OpenPGP work group at the IETF, that intends to create
a unified interface that multiple implementations provide so that
consumers can use any of those implementations interchangeably.