Rename keyring from .gpg to .pgp
This keyring contains OpenPGP certificates, and is not vendor
specific, so naming it with an extension after GnuPG in detriment
to the other multiple OpenPGP implementations does not promote
the interoperability one would expect from that ecosystem.

Given that these files are API, and will have external references,
we add a backwards compatibility symlink for now, but which should
be considered deprecated and should eventually (in the future) be
obsoleted and removed.